# api.yaml
openapi: 3.0.0
info:
    description: This is Sparta API
    version: 1.0.0
    title: Sparta_express_sequelize
    contact:
        email: flykim5115@gmail.com
servers:
    - url: "http://localhost:3000"
...
paths:
    /api/signup:
        post:
            tags:
                - user
            summary: 회원가입
            description: ㅎㅇ~
            operationId: signup
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UserRegistration"
                required: true
            responses:
                "201":
                    description: Created
                    content:
                        application/json:
                            schema:
                                properties:
                                    message:
                                        type: string
                                        example: 회원가입 성공~
                "400":
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorApiResponse"
                "412":
                    description: User already exists
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorApiResponse"
    /api/login:
        post:
            tags:
                - user
            summary: 로그인
            description: login
            requestBody:
                description: login
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UserRegistration"

            responses:
                201:
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AuthApiResponse"
                412:
                    description: Fail
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorApiResponse"
...
components:
    schemas:
        UserRegistration:
            type: object
            properties:
                nickname:
                    type: string
                    minLength: 3
                password:
                    type: string
                    minLength: 4
            required:
                - nickname
                - password
            example:
                username: kingseungil
                password: password
...
        AuthApiResponse:
            type: object
            properties:
                token:
                    type: string
                username:
                    type: string
            required:
                - token
                - username
            example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
                username: kingseungil
        ErrorApiResponse:
            type: object
            properties:
                message:
                    type: string
            required:
                - message
            example:
                message: "Something went wrong 🤪"
...
    securitySchemes:
        jwt_auth:
            description: Bearer token authorization with JWT
            type: http
            scheme: bearer
            bearerFormat: JWT